Sync l2jpd 6871, 6872, 6876, 6878, 6885, 6881, 6890, 6894, 6905, 6906, 6907, 6917, 6918, 6919, 6920
git-svn-id: http://svn.l2jfree.com/l2jfree@8729 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-datapack/sql/seven_signs_status.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS `seven_signs_status` (
   `current_cycle` INT(10) NOT NULL DEFAULT 1,
   `festival_cycle` INT(10) NOT NULL DEFAULT 1,
   `active_period` INT(10) NOT NULL DEFAULT 1,
-  `date` INT(10) NOT NULL DEFAULT 1,
+  `date` DECIMAL(20,0) NOT NULL DEFAULT 1,
   `previous_winner` INT(10) NOT NULL DEFAULT 0,
   `dawn_stone_score` DECIMAL(20,0) NOT NULL DEFAULT 0,
   `dawn_festival_score` INT(10) NOT NULL DEFAULT 0,
